requirements/base.txt


# Bleeding edge Django
django==1.10.1

# Configuration
django-environ==0.4.0



# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0

# Models
django-model-utils==2.6

# Images
Pillow==3.3.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.27.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.5.0
redis>=2.10.5


celery==3.1.23
jsonfield==1.0.3
djangorestframework==3.4.7
xmltodict==0.10.2

# Your custom requirements go here
